add PreProcessRawFile script
this script is required to convert the .raw text file from the probe to a .mat file that can be loaded into SlugPen. Run this script and when prompted, select your .raw text file. If the code break, it is because your .raw text file format is incompatible. These formats change a lot, so you'll need to update the PreProcessRawFile script to be compatible with your .raw file format. This helps us keep changes with the instrument itseful (which are common) to not affect SlugPen.